+++ app/Http/Controllers/AddressController.php
@@ -2,12 +2,12 @@
 
 namespace App\Http\Controllers;
 
-use Illuminate\Http\Request;
-use App\Address;
 use App\Helpers\Notiflix;
+use App\Models\Address;
+use Illuminate\Foundation\Testing\WithoutMiddleware;
+use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Log;
-use Illuminate\Foundation\Testing\WithoutMiddleware;
 
 class AddressController extends Controller
 {
