+++ app/Models/Order.php
@@ -1,17 +1,27 @@
 <?php
 
-namespace App;
+namespace App\Models;
 
 use Illuminate\Database\Eloquent\Model;
 
-class Cart extends Model
+class Order extends Model
 {
     /**
      * The attributes that are mass assignable.
      *
      * @var array
      */
-    protected $fillable = ['user_id', 'product_id', 'qty', 'price', 'discount'];
+    protected $fillable = [
+        'user_id',
+        'product_id',
+        'address_id',
+        'order_id',
+        'qty',
+        'price',
+        'discount',
+        'status',
+        'type',
+    ];
 
     public function user()
     {
@@ -21,6 +31,11 @@ class Cart extends Model
     public function product()
     {
         return $this->belongsTo(Product::class, 'product_id');
+    }
+
+    public function address()
+    {
+        return $this->belongsTo(Address::class, 'address_id');
     }
 
     public function getDiscountedPriceAttribute()
