+++ database/seeds/ProductSeeder.php
@@ -2,8 +2,9 @@
 
 namespace Database\Seeders;
 
-use App\FileImage;
-use App\Product;
+use App\Models\FileImage;
+use App\Models\Product;
+use Faker\Factory;
 use Illuminate\Database\Seeder;
 use Illuminate\Support\Str;
 
@@ -11,7 +12,7 @@ class ProductSeeder extends Seeder
 {
     public function __construct()
     {
-        $this->faker = Faker\Factory::create();
+        $this->faker = Factory::create();
         $this->faker->seed(1234);
     }
