+++ tests/Feature/RegisterTest.php
@@ -1,7 +1,7 @@
 <?php
 
 namespace Tests\Feature;
-use App\User;
+use App\Models\User;
 use Illuminate\Foundation\Testing\RefreshDatabase;
 use Illuminate\Foundation\Testing\WithFaker;
 use Tests\TestCase;
