+++ tests/Feature/SellerAuthTest.php
@@ -2,7 +2,7 @@
 
 namespace Tests\Feature;
 
-use App\User;
+use App\Models\User;
 use Illuminate\Support\Facades\Auth;
 
 use Tests\TestCase;
