+++ tests/Unit/AddressUnitTest.php
@@ -4,7 +4,7 @@ namespace Tests\Unit;
 
 use PHPUnit\Framework\TestCase;
 use App\Address;
-use App\User;
+use App\Models\User;
 
 class AddressUnitTest extends TestCase
 {
