+++ tests/Unit/AdminTest.php
@@ -6,7 +6,7 @@ use App\Approval;
 use Illuminate\Foundation\Testing\RefreshDatabase;
 use Illuminate\Foundation\Testing\WithFaker;
 use Illuminate\Support\Facades\Auth;
-use App\User;
+use App\Models\User;
 use Faker\Generator as Faker;
 use Illuminate\Support\Facades\Log;
 use Tests\TestCase;
