+++ tests/Unit/CartUnitTest.php
@@ -6,7 +6,7 @@ use Illuminate\Foundation\Testing\DatabaseMigrations;
 use Illuminate\Foundation\Testing\DatabaseTransactions;
 use App\Cart;
 use App\Product;
-use App\User;
+use App\Models\User;
 use PHPUnit\Framework\TestCase;
 
 class CartUnitTest extends TestCase
