+++ tests/Unit/OrderUnitTest.php
@@ -4,7 +4,7 @@ namespace Tests\Unit;
 
 use PHPUnit\Framework\TestCase;
 use App\Order;
-use App\User;
+use App\Models\User;
 use App\Product;
 use App\Address;
 class OrderUnitTest extends TestCase
