+++ tests/Unit/SellerCreateTest.php
@@ -2,7 +2,7 @@
 
 namespace Tests\Unit;
 
-use App\User;
+use App\Models\User;
 use App\Product;
 use Faker\Generator as Faker;
 use App\Cart;
