+++ tests/Unit/SellerUnitTest.php
@@ -5,7 +5,7 @@ namespace Tests\Unit;
 use PHPUnit\Framework\TestCase;
 
 use App\Seller;
-use App\User;
+use App\Models\User;
 
 class SellerUnitTest extends TestCase
 {
